pyproject.toml
[project]
name = "reflection"
version = "0.1.0"
description = "Self-evolving autonomous multi-agent system for coding problems"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.44",
    "pydantic>=2.0",
    "lancedb>=0.20",
    "sentence-transformers>=3.0",
    "duckdb>=1.0",
    "typer>=0.12",
    "structlog>=24.0",
    "tomli>=2.0;python_version<'3.11'",
    "tomli-w>=1.0",
    "python-ulid>=3.1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.28",
    "asyncssh>=2.17",
    "pyyaml>=6.0",
]

[project.scripts]
reflection = "cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["data/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.hatch.build.targets.wheel]
packages = ["agenix", "cli", "services", "tools"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
